VectorizedKmers

Documentation for VectorizedKmers.

VectorizedKmers.AbstractKmerCount

VectorizedKmers.AbstractKmerCountVector

VectorizedKmers.KmerCount

VectorizedKmers.count_kmers!

VectorizedKmers.AbstractKmerCount — Type

AbstractKmerCount{A, K, T <: Real, V <: AbstractVector{T}}

Abstract type for K-mer counts. A is the alphabet size, K is the K-mer size, and T is the element type of the underlying counts field, which in turn has type V{T}.

VectorizedKmers.AbstractKmerCountVector — Type

AbstractKmerCountVector{A, K, T <: Real, M <: AbstractMatrix{T}} <: AbstractVector{KmerCount{A, K, T}}

A container for K-mer counts, where K-mer counts are stored together as rows or columns in a matrix. A is the alphabet size, K is the K-mer size, T is the element type of the counts, and M is the type of the matrix in which the K-mer counts are stored.

VectorizedKmers.KmerCount — Type

KmerCount{A, K, T, V} <: AbstractKmerCount{A, K, T, V}

A concrete type for K-mer counts with vector type Base.Vector{T}.

VectorizedKmers.count_kmers! — Method

count_kmers!(kmer_count, kmers; reset=true)

Mutate the counts vector in kmer_count by adding the counts of each kmer in kmers. The K-mers in kmers must be represented as integers between 0 and length(kmer_count) - 1.

If reset is true, the counts vector will be zero-ed before counting.

This is not a very efficient method, since it takes an entire vector. It is mainly used for testing. Ideally the K-mers would be procedurally calculated in constant memory.