API Reference

VectorizedKmers.AbstractKmerArray — Type

AbstractKmerArray{N, S, k, T <: Real, A <: AbstractArray{T, N}} <: AbstractArray{T, N}

Abstract type representing k-mer count in N-dimensional arrays. This type serves as the foundation for various concrete k-mer count structures.

Type Parameters

N: Dimensionality of the array.

S: Alphabet size.

k: k-mer size.

T: Element type, subtype of Real.

A: Array type, subtype of AbstractArray{T, N}.

Rationale for Type Parameter Ordering

The order of the type parameters is purposefully chosen based on the logical importance of comparisons:

N: Dimensionality is paramount. Types with differing dimensions cannot be compared directly. They would somehow need to be transformed to the same dimension.

S: Alphabet size is a crucial differentiator. It's nonsensical to compare types with distinct alphabet sizes, even if k-mer sizes and other attributes are the same.

k: k-mer size is the next important type parameter. Types with varying k-mer sizes shouldn't be directly compared even if they share other attributes.

T: Element type might imply certain transformations or states (e.g., normalization when using floats). Types with different element types can have distinct implications and hence shouldn't be equated.

A: Lastly, the specific array type (like Vector or SparseVector) is the most flexible comparison point. Types might be comparable if they share all other parameters, even if their array types differ.

This order ensures that the most significant distinctions between types are made first, allowing for more meaningful and nuanced comparisons.

VectorizedKmers.KmerVector — Type

KmerVector{S, k, T, V} <: AbstractKmerVector{S, k, T, V}

A is the alphabet size, k is the k-mer size, and T is the element type of the underlying values field, which in turn has type V.

VectorizedKmers.KmerVectors — Type

KmerVectors

A container for k-mer vectors, where the values are stored together as columns or rows in a matrix.

Note

It might be more efficient to store the values as columns in a matrix, since the elements in a column are contiguous in memory.

D is the dimension of the matrix in which the vectors are next to each other (1 for rows, 2 for columns). S is the alphabet size, k is the k-mer length, and T is the element type of the underlying values field, which in turn has type V.

VectorizedKmers.count_kmers — Function

count_kmers(sequence::Vector{<:Integer}, S, k; zeros=zeros, reset=true)

This is the default k-mer counting method. It creates a new S^k-sized vector using zeros and counts the k-mers in sequence. Since S is the alphabet, and the elements in sequence are integers, the maximum value in sequence must be less than S

VectorizedKmers.count_kmers! — Function

count_kmers!(kmer_vector, sequence)

Mutates kmer_vector.

If reset is true, the array will be zero-ed before counting.