API Reference

VectorizedKmers.AbstractKmerArray — Type

AbstractKmerArray{N, S, k, T, A <: AbstractArray{T, N}} <: AbstractArray{T, N}

Abstract type representing k-mer count in N-dimensional arrays. This type serves as the foundation for various concrete k-mer count structures.

Type Parameters

N: Dimensionality of the array

S: Alphabet size

k: k-mer size

T: Element type

A: Array type, subtype of AbstractArray{T, N}

VectorizedKmers.KmerVector — Type

KmerVector{S, k, T, V} <: AbstractKmerVector{S, k, T, V}

A is the alphabet size, k is the k-mer size, and T is the element type of the underlying values field, which in turn has type V.

VectorizedKmers.KmerVectors — Type

KmerVectors

A container for k-mer vectors, where the values are stored together as columns or rows in a matrix.

D is the dimension of the vectors (i.e. 1 for column-vectors, 2 for row-vectors). S is the alphabet size, k is the k-mer length, and T is the element type of the underlying values field, which in turn has type M.

VectorizedKmers.count_kmers — Function

count_kmers(sequence, S, k; T=Int, zeros=zeros)

Creates a new S^k-sized vector using the supplied zeros function and counts the k-mers in sequence. Since S is the alphabet, and the elements in sequence are integers, the maximum value in sequence must be less than S

VectorizedKmers.count_kmers! — Function

count_kmers!(kmer_vector, sequence; reset=true)

Mutates kmer_vector.

If reset is true, the array will be zero-ed before counting.