API Reference

VectorizedKmers.AbstractKmerCountMatrix — Type

AbstractKmerCountMatrix{S, k, T <: Real, M <: AbstractMatrix{T}} <: AbstractVector{KmerCountVector{S, k, T, V} where {V <: AbstractVector{T}}}

A container for k-mer counts, where k-mer counts are stored together as rows or columns in a matrix. A is the alphabet size, k is the k-mer size, T is the element type of the counts, and M is the type of the matrix in which the k-mer counts are stored.

VectorizedKmers.AbstractKmerCounts — Type

AbstractKmerCounts

type parameters are in order of importance: N, S, k, T, A rationale: even if S, k, T, A are the same, it doesn't make sense to directly compare types with different dimension N even if k, T, A are the same, it doesn't make sense to directly compare types with different alphabet size S even if T, A are the same, it doesn't make sense to directly compare types with different k-mer size k even if array type can't be the same if T is different, it doesn't make sense to directly compare types with different element type T because that might have some meaning like having been normalized if it's float or something lastly: it does make sense to compare types with different array type A, like Vector and SparseVector, if all other parameters are the same

VectorizedKmers.KmerCountColumns — Type

KmerCountColumns{S, k, T, M} <: AbstractKmerCountMatrix{S, k, T, M}

A container for k-mer counts, where k-mer counts are stored together as columns in a matrix. This is more efficient than storing k-mer counts as rows in a matrix, since the elements in a column are contiguous in memory.

VectorizedKmers.KmerCountRows — Type

KmerCountRows{S, k, T, M} <: AbstractKmerCountMatrix{S, k, T, M}

A container for k-mer counts, where k-mer counts are stored together as rows in a matrix. This is not as efficient as storing k-mer counts as columns in a matrix, since the elements in a row are not contiguous in memory.

VectorizedKmers.KmerCountVector — Type

KmerCountVector{S, k, T, V}

A is the alphabet size, k is the k-mer size, and T is the element type of the underlying counts field, which in turn has type V.

VectorizedKmers.count_kmers — Function

count_kmers(sequence, k; zeros_func=zeros, reset=true)

Create a new S^k sized vector using zeros_func and count the k-mers in kmers.

VectorizedKmers.count_kmers! — Function

count_kmers!(kmer_count, sequence)

Mutates kmer_count.

If reset is true, the array will be zero-ed before counting.

VectorizedKmers.count_kmers_gpu — Function

count_kmers_gpu(sequences, k)

Count the k-mers in sequences using the GPU.